requirements.txt
opencv-python>=4.5.0
torch>=1.12.0
torchvision>=0.13.0
numpy>=1.21.0
pillow>=8.0.0
requests>=2.25.0
mediapipe>=0.10.0